Extract the favourite icon button
